Change one int to qint64 in QDateTime::setMSecsSinceEpoch(qint64)
This one was missed when the QDate range was extended.

+++ src/corelib/tools/qdatetime.cpp
@@ -2418,7 +2418,7 @@ void QDateTime::setMSecsSinceEpoch(qint64 msecs)
 
     QDateTimePrivate::Spec oldSpec = d->spec;
 
-    int ddays = msecs / MSECS_PER_DAY;
+    qint64 ddays = msecs / MSECS_PER_DAY;
     msecs %= MSECS_PER_DAY;
     if (msecs < 0) {
         // negative
